requirements.txt
# cli
IPy
PyGithub
hvac
openpyxl
pandas
python_graphql_client
pyyaml
requests
